# docs/swagger.yaml
basePath: /
definitions:
...
  dto.EnterPinBody:
    properties:
      pin:
        type: string
    required:
    - pin
    type: object
  dto.ForgotPasswordBody:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  dto.HistoryListResponse:
    properties:
      meta:
        $ref: '#/definitions/dto.PaginationMeta'
      transactions:
        items:
          $ref: '#/definitions/dto.HistoryResponse'
        type: array
    type: object
  dto.HistoryResponse:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: integer
      notes:
        type: string
      sender_info:
        $ref: '#/definitions/dto.SenderInfo'
      status:
        type: string
      type:
        type: string
    type: object
  dto.LoginBody:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      has_pin:
        type: boolean
      token:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  dto.PaginationMeta:
    properties:
      current_page:
        type: integer
      limit:
        type: integer
      total_data:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.RegisterBody:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  dto.ResetPasswordBody:
    properties:
      email:
        type: string
      new_password:
        minLength: 6
        type: string
    required:
    - email
    - new_password
    type: object
  dto.SenderInfo:
    properties:
      fullname:
        type: string
      phone_number:
        type: string
      wallet_id:
        type: integer
    type: object
...
  dto.SwaggerHistoryResponse:
    properties:
      data:
        $ref: '#/definitions/dto.HistoryListResponse'
      message:
        type: string
      success:
        type: boolean
    type: object
  dto.SwaggerLoginResponse:
    properties:
      data:
        $ref: '#/definitions/dto.LoginResponse'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  dto.SwaggerTopupResponse:
    properties:
      data:
        $ref: '#/definitions/dto.TopupResponse'
      message:
        type: string
      success:
        type: boolean
    type: object
  dto.SwaggerTransferResponse:
    properties:
      data:
        $ref: '#/definitions/dto.TransferResponse'
      message:
        type: string
      success:
        type: boolean
    type: object
  dto.TopupBody:
    properties:
      order_amount:
        type: number
      payment_method_id:
        type: integer
    required:
    - order_amount
    - payment_method_id
    type: object
  dto.TopupResponse:
    properties:
      created_at:
        type: string
      order_amount:
        type: number
      payment_method:
        type: string
      status:
        type: string
      tax_amount:
        type: number
      total_amount:
        type: number
      transaction_id:
        type: integer
    type: object
  dto.TransferBody:
    properties:
      amount:
        type: number
      notes:
        type: string
      pin:
        type: string
      receiver_wallet_id:
        type: integer
    required:
    - amount
    - pin
    - receiver_wallet_id
    type: object
  dto.TransferResponse:
    properties:
      amount:
        type: number
      created_at:
        type: string
      notes:
        type: string
      receiver_wallet_id:
        type: integer
      status:
        type: string
      transaction_id:
        type: integer
    type: object
...
  dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      fullname:
        type: string
      id:
        type: integer
      phone_number:
        type: string
      photo_path:
        type: string
    type: object
  pkg.BaseResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  pkg.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:9000
info:
  contact: {}
  description: Backend created by Koda using Gin
  license:
    name: MIT
  title: Backend E-WALLET
  version: "1.0"
paths:
  /auth/enter-pin:
    post:
      consumes:
      - application/json
      description: 'Set PIN setelah login pertama kali (has_pin: false)'
      parameters:
      - description: Enter PIN Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.EnterPinBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Set PIN pertama kali
      tags:
      - Auth
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Step 1 forgot password - verifikasi apakah email terdaftar
      parameters:
      - description: Email Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      summary: Verifikasi email forgot password
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login dengan email dan password, mendapatkan JWT token
      parameters:
      - description: Login Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginBody'
      produces:
      - application/json
      responses:
        "200":
          description: Login berhasil
          schema:
            $ref: '#/definitions/dto.SwaggerLoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      summary: Login user
      tags:
      - Auth
  /auth/logout:
    delete:
      description: Logout dan hapus token dari whitelist
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Membuat akun user baru dan wallet dengan saldo 0
      parameters:
      - description: Register Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/pkg.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      summary: Register user baru
      tags:
      - Auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Step 2 forgot password - ganti password dengan yang baru
      parameters:
      - description: Reset Password Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      summary: Reset password baru
      tags:
      - Auth
  /transactions/history:
    get:
      description: Mengambil riwayat transaksi user dengan filter dan pagination
      parameters:
      - description: 'Filter tipe: topup / transfer'
        in: query
        name: type
        type: string
      - description: 'Halaman (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Jumlah data (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SwaggerHistoryResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      security:
      - BearerAuth: []
      summary: History transaksi
      tags:
      - Transactions
  /transactions/topup:
    post:
      consumes:
      - application/json
      description: Menambah saldo wallet user
      parameters:
      - description: Topup Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.TopupBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SwaggerTopupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Top up saldo
      tags:
      - Transactions
  /transactions/transfer:
    post:
      consumes:
      - application/json
      description: Transfer saldo ke user lain dengan verifikasi PIN
      parameters:
      - description: Transfer Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.TransferBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SwaggerTransferResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Transfer saldo
      tags:
      - Transactions
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"