# internal/controller/auth_controller.go
package controller

import (
	"ewallet-backend/internal/dto"
	"ewallet-backend/internal/service"
	"ewallet-backend/pkg"
	"log"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/gin-gonic/gin/binding"
)

type AuthController struct {
	authService *service.AuthService
}

func NewAuthController(authService *service.AuthService) *AuthController {
	return &AuthController{authService: authService}
}

// Register godoc
//
//	@Summary		Register user baru
//	@Description	Membuat akun user baru dan wallet dengan saldo 0
//	@Tags			Auth
//	@Accept			json
//	@Produce		json
//	@Param			body	body		dto.RegisterBody	true	"Register Request"
//	@Success		201		{object}	pkg.BaseResponse
//	@Failure		400		{object}	pkg.ErrorResponse
//	@Router			/auth/register [post]
func (a *AuthController) Register(ctx *gin.Context) {
	var body dto.RegisterBody
	if err := ctx.ShouldBindWith(&body, binding.JSON); err != nil {
		log.Println("Error:", err.Error())
		ctx.JSON(http.StatusBadRequest, pkg.ErrorResponse{
			Message: "Bad Request",
			Success: false,
			Error:   err.Error(),
		})
		return
	}

	if err := a.authService.Register(ctx.Request.Context(), body); err != nil {
		log.Println("Error:", err.Error())
		ctx.JSON(http.StatusConflict, pkg.ErrorResponse{
			Message: "Error",
			Success: false,
			Error:   err.Error(),
		})
		return
	}

	ctx.JSON(http.StatusCreated, pkg.BaseResponse{
		Message: "Registration successful",
		Success: true,
	})
}

// Login godoc
//
//	@Summary		Login user
//	@Description	Login dengan email dan password, mendapatkan JWT token
//	@Tags			Auth
//	@Accept			json
//	@Produce		json
//	@Param			body	body		dto.LoginBody				true	"Login Request"
//	@Success		200		{object}	dto.SwaggerLoginResponse	"Login berhasil"
//	@Failure		400		{object}	pkg.ErrorResponse
//	@Failure		401		{object}	pkg.ErrorResponse
//	@Router			/auth/login [post]
func (a *AuthController) Login(ctx *gin.Context) {
	var body dto.LoginBody
	if err := ctx.ShouldBindWith(&body, binding.JSON); err != nil {
		log.Println("Error:", err.Error())
		ctx.JSON(http.StatusBadRequest, pkg.ErrorResponse{
			Message: "Bad Request",
			Success: false,
			Error:   err.Error(),
		})
		return
	}

	result, err := a.authService.Login(ctx.Request.Context(), body)
	if err != nil {
		log.Println("Error:", err.Error())
		ctx.JSON(http.StatusUnauthorized, pkg.ErrorResponse{
			Message: "Unauthorized",
			Success: false,
			Error:   err.Error(),
		})
		return
	}

	ctx.JSON(http.StatusOK, pkg.Response[dto.LoginResponse]{
		Message: "Login successful",
		Success: true,
		Data:    result,
	})
}

// Logout godoc
//
//	@Summary		Logout user
//	@Description	Logout dan hapus token dari whitelist
//	@Tags			Auth
//	@Produce		json
//	@Success		200	{object}	pkg.BaseResponse
//	@Failure		401	{object}	pkg.ErrorResponse
//	@Failure		500	{object}	pkg.ErrorResponse
//	@Security		BearerAuth
//	@Router			/auth/logout [delete]
func (a *AuthController) Logout(ctx *gin.Context) {
	token, _ := ctx.Get("token")

	if err := a.authService.Logout(ctx.Request.Context(), token.(string)); err != nil {
		log.Println("Error:", err.Error())
		ctx.JSON(http.StatusInternalServerError, pkg.ErrorResponse{
			Message: "Internal Error",
			Success: false,
			Error:   "internal server error",
		})
		return
	}

	ctx.JSON(http.StatusOK, pkg.BaseResponse{
		Message: "Logout successful",
		Success: true,
	})
}

// ForgotPassword godoc
//
//	@Summary		Verifikasi email forgot password
//	@Description	Step 1 forgot password - verifikasi apakah email terdaftar
//	@Tags			Auth
//	@Accept			json
//	@Produce		json
//	@Param			body	body		dto.ForgotPasswordBody	true	"Email Request"
//	@Success		200		{object}	pkg.BaseResponse
//	@Failure		400		{object}	pkg.ErrorResponse
//	@Failure		404		{object}	pkg.ErrorResponse
//	@Router			/auth/forgot-password [post]
func (a *AuthController) ForgotPassword(ctx *gin.Context) {
	var body dto.ForgotPasswordBody
	if err := ctx.ShouldBindWith(&body, binding.JSON); err != nil {
		log.Println("Error:", err.Error())
		ctx.JSON(http.StatusBadRequest, pkg.ErrorResponse{
			Message: "Bad Request",
			Success: false,
			Error:   err.Error(),
		})
		return
	}

	if err := a.authService.ForgotPassword(ctx.Request.Context(), body); err != nil {
		log.Println("Error:", err.Error())

		// Tentukan status code
		statusCode := http.StatusInternalServerError
		if err.Error() == "email not found" {
			statusCode = http.StatusNotFound
		}

		ctx.JSON(statusCode, pkg.ErrorResponse{
			Message: "Failed",
			Success: false,
			Error:   err.Error(),
		})
		return
	}

	ctx.JSON(http.StatusOK, pkg.BaseResponse{
		Message: "Email verified. You can now reset your password.",
		Success: true,
	})
}

// ResetPassword godoc
//
//	@Summary		Reset password baru
//	@Description	Step 2 forgot password - ganti password dengan yang baru
//	@Tags			Auth
//	@Accept			json
//	@Produce		json
//	@Param			body	body		dto.ResetPasswordBody	true	"Reset Password Request"
//	@Success		200		{object}	pkg.BaseResponse
//	@Failure		400		{object}	pkg.ErrorResponse
//	@Failure		404		{object}	pkg.ErrorResponse
//	@Failure		500		{object}	pkg.ErrorResponse
//	@Router			/auth/reset-password [post]
func (a *AuthController) ResetPassword(ctx *gin.Context) {
	var body dto.ResetPasswordBody
	if err := ctx.ShouldBindWith(&body, binding.JSON); err != nil {
		log.Println("Error:", err.Error())
		ctx.JSON(http.StatusBadRequest, pkg.ErrorResponse{
			Message: "Bad Request",
			Success: false,
			Error:   err.Error(),
		})
		return
	}

	if err := a.authService.ResetPassword(ctx.Request.Context(), body); err != nil {
		log.Println("Error:", err.Error())

		statusCode := http.StatusInternalServerError
		if err.Error() == "email not found" {
			statusCode = http.StatusNotFound
		}

		ctx.JSON(statusCode, pkg.ErrorResponse{
			Message: "Failed",
			Success: false,
			Error:   err.Error(),
		})
		return
	}

	ctx.JSON(http.StatusOK, pkg.BaseResponse{
		Message: "Password reset successfully. Please login with your new password.",
		Success: true,
	})
}

// EnterPin godoc
//
//	@Summary		Set PIN pertama kali
//	@Description	Set PIN setelah login pertama kali (has_pin: false)
//	@Tags			Auth
//	@Accept			json
//	@Produce		json
//	@Param			body	body		dto.EnterPinBody	true	"Enter PIN Request"
//	@Success		200		{object}	pkg.BaseResponse
//	@Failure		400		{object}	pkg.ErrorResponse
//	@Failure		401		{object}	pkg.ErrorResponse
//	@Security		BearerAuth
//	@Router			/auth/enter-pin [post]
func (a *AuthController) EnterPin(ctx *gin.Context) {
	// Ambil claims dari token (sudah login)
	token, _ := ctx.Get("claims")
	claims := token.(pkg.Claims)

	var body dto.EnterPinBody
	if err := ctx.ShouldBindWith(&body, binding.JSON); err != nil {
		log.Println("Error:", err.Error())
		ctx.JSON(http.StatusBadRequest, pkg.ErrorResponse{
			Message: "Bad Request",
			Success: false,
			Error:   err.Error(),
		})
		return
	}

	if err := a.authService.EnterPin(ctx.Request.Context(), claims.Id, body); err != nil {
		log.Println("Error:", err.Error())

		statusCode := http.StatusBadRequest
		if err.Error() == "pin already set. use change pin to update" {
			statusCode = http.StatusConflict // 409
		}

		ctx.JSON(statusCode, pkg.ErrorResponse{
			Message: "Failed",
			Success: false,
			Error:   err.Error(),
		})
		return
	}

	ctx.JSON(http.StatusOK, pkg.BaseResponse{
		Message: "PIN has been set successfully",
		Success: true,
	})
}